+++ builder/src/attribute/bonuses/weapon_attribute.rs
@@ -1,5 +1,6 @@
 use std::fmt::Display;
 
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 use crate::{
@@ -33,7 +34,7 @@ impl From<(WeaponHand, WeaponStat)> for Attribute {
 }
 
 impl GetBonuses for WeaponAttribute {
-    fn get_bonuses(&self, value: f32) -> Option<Vec<Bonus>> {
+    fn get_bonuses(&self, value: Decimal) -> Option<Vec<Bonus>> {
         match self {
             Self(hand, WeaponStat::CriticalMultiplier) => Some(vec![Bonus::new(
                 (*hand, WeaponStat::CriticalMultiplier1920).into(),
